applications/speech_recognition/pyproject.toml
[tool.poetry]
name = "speech-recognition"
version = "0.1.0"
description = ""
authors = ["Jacek Lipiec <jacek.lipiec.bc@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.10"
openai-whisper = "20240930"
torch = { version = "2.4.1", source = "pytorch-gpu" }
torchaudio = { version = "2.4.1", source = "pytorch-gpu" }
numpy = "^2.0.2"
Flask = "^3.0.3"
python-config = {path = "../../libraries/python_config", develop = true}
python-utilities = {path = "../../libraries/python_utilities", develop = true}
python-sound-input = {path = "../../libraries/python_sound_input", develop = true}

[[tool.poetry.packages]]
include = "speech_recognition"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[tool.poetry.scripts]
serve = "speech_recognition.server:serve"
listen = "speech_recognition.cli:listen"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
